+++ src/actions/todoListActions.ts
@@ -1,0 +1,7 @@
+import * as actionTypes from './types/todolistTypes';
+import { ISingleElementList } from '../entitiens/todoSingleEl';
+
+export const setNewElemTodoList = (newElem: ISingleElementList) => ({
+  type: actionTypes.SET_NEW_ELEM,
+  newElem,
+});