+++ CHANGELOG.md
@@ -5,11 +5,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.2.0] - 2020-05-02
+## [1.4.0] - 2020-05-16
 
 ### Changes
 
     -
+
+### Added
+
+    - new components Form and TodoListComp
+    - new plik entities (interface global)
+    - todoListReducer
+
+## [1.3.0] - 2020-05-04
+
+### Changes
+
+    - ThirdScreen.tsx
+
+### Added
+
+    - adding text and graphics in the file ThirdScreen
+
+## [1.2.0] - 2020-05-02
+
+### Changes
+
+    - Main.tsx
 
 ### Added
 
@@ -19,22 +41,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
-    - new logo splashscreen
-    - backgrand splashscreen
+- app.json
 
 ### Added
 
-    - splashscreen
+    - new logo splashscreen
+    - backgrand splashscreen
 
 ## [1.0.1] - 2020-04-15
 
 ### Changes
 
-    -
+    - App.tsx
 
 ### Added
 
     - add icon
+    - navigation
 
 ## [1.0.0] - 2020-03-24